refactor: rename with simpler file name
